requirements.txt
pydantic==2.8.2
pydantic-settings==2.3.4
aiokafka==0.11.0
python-dotenv==1.0.1
loguru==0.7.2
clickhouse-driver==0.2.8
